+++ src/contact/Contact.scss
@@ -1,175 +1,8 @@
-.mi-formulario {
-    opacity: 0.4;
-    transition: opacity 1s;
-  }
-  
-
-
-.formulario_contenedor{
-    font-family: 'FormularioFont', Arial, sans-serif;
-    background-color: #333;
-    padding: 20%;
-    width: 180%;
-    opacity: 0.4;
-    transition: opacity 1s;
-    height: 90%;
-    border-radius: 1rem;
-    margin-left: 130%;
-    margin-top: 20%;
-   box-shadow: 0px 0px 10px 1px white;
-
-
-}
-.mi-formulario-entrando {
-    opacity: 1;
-  }
-
-.formulario_espacio{
-
+.contenedor_de_formulario{
     width: 100%;
+    margin-top: 2%;
+}
+.formulario_contacto{
+    overflow-y: scroll;
 
-}
-.formulario_inputs{
-    padding: 12%;
-    border: none;
-    border-radius: 0.5rem;
-    background-color: #282828;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-    margin-bottom: 5%;
-}
-.formulario_errores{
-    padding: 0%;
-    margin-top: 0.3%;
-    color: red;
-    background: none;
-    font-size: 80%;
-    position: absolute;
-}
-.formulario_errores_apellido{
-    padding: 0%;
-    color: red;
-    margin-top: 0.3%;
-    background: none;
-    margin-left: 5.5%;
-    font-size: 80%;
-    position: absolute;
-}
-.formulario_nombres{
-    display: flex;
-}
-.formulario_apellido{
-    padding: 10%;
-    border: none;
-    text-align: center;
-    border-radius: 0.5rem;
-    color: white;
-    background-color: #282828;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-    margin-left: 50%;
-    margin-bottom: 5%;
-    outline: none;
-}
-.formulario_nombre{
-    padding: 10%;
-    border: none;
-    text-align: center;
-    color:white;
-    border-radius: 0.5rem;
-    background-color: #282828;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-    margin-bottom: 5%;
-    outline: none;
-}
-.formulario_label{
-    font-size: 100%;
-    font-weight: 600;
-    color: #333;
-    margin-bottom: 5%;
-}
-.formulario_inputs_correo{
-    padding: 5%;
-    text-align: center;
-    width: 100%;
-    outline: none;
-    border: none;
-    color: white;
-    border-radius: 0.5rem;
-    background-color: #282828;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-}
-.formulario_letras_apellido{
-    color: white;
-    margin-left: 90%;
-}
-.formulario_letras_mensaje{
-    color: white;
-    margin-left: 42%;
-    margin-top: 10%;
-}
-.formulario_inputs_textArea{
-    width: 100%;
-    padding: 5%;
-    height: 40%;
-    border: none;
-    outline: none;
-    color: white;
-    border-radius: 0.5rem;
-    background-color: #282828;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-    resize: none;
-}
-.formulario-submit{
-    background-color: transparent;
-    color: #fff;
-    padding: 2%;
-    border-radius: 4%;
-    border:none;
-    margin-top: 5%;
-    font-size: 100%;
-    font-weight: 600;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-    cursor: pointer;
-    &:hover{
-        background-color: #1DB954;
-        -webkit-box-shadow: 0px 0px 15px 5px #1DB954;
-    }
-}
-.formulario_letras{
-    color: white;
-}
-.formulario_letras_nombre{
-    color: white;
-    margin-left: 40%;
-}
-.formulario_letras_correo{
-    color: white;
-    margin-left: 43%;
-    margin-top: -3%;
-}
-.formulario-message {
-    font-size: 14px;
-    font-weight: 600;
-    visibility: hidden;
-  }
-.formulario_enviar{
-    color: white;
-}
-.formulario-message{
-    color: red;
-}
-.formulario_correcto{
-    color: #1DB954;
-}
-.formulario_titulo{
-    font-size: 170%;
-    font-weight: 600;
-    margin-top: -10%;
-    margin-left: 30%;
-    margin-bottom: 5%;
 }
